+++ src/client.ts
@@ -3,6 +3,10 @@ import { Transport } from "./transport";
 import { BluetickError } from "./errors";
 import { VERSION } from "./version";
 import { AccountResource } from "./resources/account";
+import { MessagesResource } from "./resources/messages";
+import { WebhooksResource } from "./resources/webhooks";
+import { AudiencesResource } from "./resources/audiences";
+import { CampaignsResource } from "./resources/campaigns";
 import type { Ping } from "./types/ping";
 
 const DEFAULT_BASE_URL = "https://api.blueticks.co";
@@ -20,6 +24,10 @@ export interface BluetickOptions {
 
 export class Blueticks {
   readonly account: AccountResource;
+  readonly messages: MessagesResource;
+  readonly webhooks: WebhooksResource;
+  readonly audiences: AudiencesResource;
+  readonly campaigns: CampaignsResource;
   private readonly transport: Transport;
   readonly baseUrl: string;
 
@@ -46,6 +54,10 @@ export class Blueticks {
     });
 
     this.account = new AccountResource(this);
+    this.messages = new MessagesResource(this);
+    this.webhooks = new WebhooksResource(this);
+    this.audiences = new AudiencesResource(this);
+    this.campaigns = new CampaignsResource(this);
   }
 
   async ping(opts: { signal?: AbortSignal } = {}): Promise<Ping> {