+++ src/resources/index.ts
@@ -1,2 +1,16 @@
 export { AccountResource } from "./account";
 export { PingResource } from "./ping";
+export { MessagesResource, type SendMessageParams } from "./messages";
+export {
+  WebhooksResource,
+  type CreateWebhookParams,
+  type UpdateWebhookParams,
+} from "./webhooks";
+export {
+  AudiencesResource,
+  type ContactInput,
+  type CreateAudienceParams,
+  type UpdateAudienceParams,
+  type UpdateContactParams,
+} from "./audiences";
+export { CampaignsResource, type CreateCampaignParams } from "./campaigns";