+++ src/types/index.ts
@@ -1,2 +1,34 @@
 export { AccountSchema, type Account } from "./account";
 export { PingSchema, type Ping } from "./ping";
+export {
+  MessageSchema,
+  MessageStatusSchema,
+  type Message,
+  type MessageStatus,
+} from "./messages";
+export {
+  WebhookSchema,
+  WebhookStatusSchema,
+  WebhookEventTypeSchema,
+  WebhookCreateResultSchema,
+  WebhookEventSchema,
+  type Webhook,
+  type WebhookStatus,
+  type WebhookEventType,
+  type WebhookCreateResult,
+  type WebhookEvent,
+} from "./webhooks";
+export {
+  AudienceSchema,
+  ContactSchema,
+  AppendContactsResultSchema,
+  type Audience,
+  type Contact,
+  type AppendContactsResult,
+} from "./audiences";
+export {
+  CampaignSchema,
+  CampaignStatusSchema,
+  type Campaign,
+  type CampaignStatus,
+} from "./campaigns";